chore(ui): move store file to hook
