+++ .env.example
@@ -1,0 +1,17 @@
+#API
+APP_ENV=develop
+API_URL=
+API_PORT=3001
+
+API_PREFIX=
+
+JWT_SECRET=
+JWT_EXPIRATION_TIME=1h
+
+FILE_UPLOAD=_uploads
+FILE_PATH=
+
+DATABASE_URL=postgresql://USER:PASSWORD@HOST:PORT/DATABASE
+
+# UI
+NEXT_PUBLIC_URL=
